fix bugs

+++ inference/clip_infer.py
@@ -6,9 +6,11 @@ import shutil
 from modelscope.hub.utils.utils import get_cache_dir
 
 
-def clip_infer(input_image):
-    cache_dir = get_cache_dir()
-    shutil.rmtree(cache_dir)
+def clip_infer(input_image, clear_cache=False):
+    if clear_cache:
+        print("Clear download model weights.")
+        cache_dir = get_cache_dir()
+        shutil.rmtree(cache_dir)
     model_id = 'damo/cv_clip-interrogator'
     pipe = pipeline(Tasks.image_captioning, model=model_id, model_revision='v1.0.0')
     picture = pipe(input_image)['caption']
